Streamline dev-test stack

+++ spec/rails_helper.rb
@@ -104,7 +104,7 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = false
 
   config.before :suite do
-    WebMock.disable_net_connect!(allow: ['localhost', '127.0.0.1', 'fedora', 'solr', 'matterhorn', 'https://chromedriver.storage.googleapis.com'])
+    WebMock.disable_net_connect!(allow: ['localhost', '127.0.0.1', 'fedora', 'fedora-test', 'solr', 'solr-test', 'matterhorn', 'https://chromedriver.storage.googleapis.com'])
     DatabaseCleaner.clean_with(:truncation)
     ActiveFedora::Cleaner.clean!
     disable_production_minter!
@@ -119,7 +119,7 @@ RSpec.configure do |config|
   end
 
   config.after :suite do
-    if Settings.spec['fake_dropbox']
+    if Settings.spec && Settings.spec['fake_dropbox']
       FileUtils.remove_dir Settings.spec['fake_dropbox'], true
       Settings.dropbox.path = Settings.spec['real_dropbox']
       Settings.spec = nil
